pyproject.toml
[project]
name = "ecosystem-narrator"
version = "0.1.0"
description = "Gemini-Powered Agricultural Ecosystem Narration System"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Ruchit Agrawal", email = "rragrawal16@gmail.com" }]
license = { text = "MIT" }

dependencies = [
    "google-genai>=1.0",
    "pydantic>=2.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "rich>=13.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
ecosystem-narrator = "ecosystem_narrator.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
strict = true
python_version = "3.11"
